All simulation parameters are passed via command line arguments now (update).

+++ Dash2/github/initial_state_loader.py
@@ -5,7 +5,7 @@ import time
 import numpy as np
 import cPickle as pickle
 from distributed_event_log_utils import load_id_dictionary, collect_unique_user_event_pairs
-from user_repo_graph_utils import build_graph_from_csv, partition_graph, print_user_profiles
+from user_repo_graph_utils import build_graph_from_csv, partition_graph, print_user_profiles, print_graph, print_users_neighborhood_sizes
 from distributed_event_log_utils import event_types
 
 '''
@@ -40,6 +40,10 @@ def build_state_from_event_log(input_event_log, number_of_user_partitions=1, sta
 
     print "printing graph..."
     print_user_profiles(G, input_event_log + "_users.json", number_of_user_partitions)
+    graph_file_name = input_event_log + "_UR_graph.pickle"
+    print_graph(G, graph_file_name)
+    users_neighborhood_sizes_file = input_event_log + "_users_neighborhood_sizes.pickle"
+    print_users_neighborhood_sizes(G, users_neighborhood_sizes_file)
 
     users_file = input_event_log + "_users.json"
     repos_file = input_event_log + "_repos.json"
@@ -58,8 +62,9 @@ def build_state_from_event_log(input_event_log, number_of_user_partitions=1, sta
             "users_ids": users_ids,
             "repos_ids": repos_ids,
             "number_of_partitions": number_of_user_partitions,
-            "embedding_files": embedding_files,
-            "event_rate_model_file": ""
+            "event_rate_model_file": "",
+            "UR_graph_path": graph_file_name,
+            "users_neighborhood_sizes_file": users_neighborhood_sizes_file
         }
     }
     state_file_name = input_event_log + "_state.json" if state_file_name is None else state_file_name
@@ -190,32 +195,23 @@ class EmbeddingCalculator(object):
 Populates agent's decision data objects with repo probabilities. Probabilities either loaded from a .prob file 
 (precomputed values of probability vectors) or from embedding files.
 '''
-def populate_embedding_probabilities(agents_decision_data, state_file, probability_vector_size = 10):
-    initial_state_meta_data = read_state_file(state_file)
-    if initial_state_meta_data["embedding_files"] is not None and initial_state_meta_data["embedding_files"] != "":
-        embedding_files_data = initial_state_meta_data["embedding_files"]
-        UsimId2strId = load_id_dictionary(initial_state_meta_data["users_ids"], isSimId2strId=True)
-        strId2RsimId = load_id_dictionary(initial_state_meta_data["repos_ids"], isSimId2strId=False)
-        for event_type, embedding_info in embedding_files_data.iteritems():
-            if "probabilities_file" in embedding_info:
-                print "Event type: ", event_type, ", loading probabilities from .prob files (.prob - probabilities are precomputed in advance)."
-                prob_file = open(embedding_info["probabilities_file"], 'rb')
-                probabilities = pickle.load(prob_file); prob_file.close()
-                for _, decision_data in agents_decision_data.iteritems():
-                    if decision_data.id in probabilities:
-                        decision_data.embedding_probabilities[event_type] = probabilities[decision_data.id]
-            else: # this takes a long time, so in normal scenario it should not hit this case (assuming probabilities are precomputed)
-                print "Event type: ", event_type, ", computing probabilities from embedding. May take a while... "
-                calculator = EmbeddingCalculator(embedding_info["file_name"], embedding_info["dictionary"], UsimId2strId, strId2RsimId)
-                for agent_id, decision_data in agents_decision_data.iteritems():
-                    decision_data.embedding_probabilities[event_type] = calculator.calculate_probabilities(decision_data.id, probability_vector_size)
+def populate_embedding_probabilities(agents_decision_data, embedding_path):
+    for event_type in event_types:
+        prob_file_path = str(embedding_path) + event_type + ".prob"
+        if os.path.isfile(prob_file_path):
+            print "Event type: ", event_type, ", loading probabilities from .prob files (.prob - probabilities are precomputed in advance)."
+            prob_file = open(prob_file_path, 'rb')
+            probabilities = pickle.load(prob_file); prob_file.close()
+            for _, decision_data in agents_decision_data.iteritems():
+                if decision_data.id in probabilities:
+                    decision_data.embedding_probabilities[event_type] = probabilities[decision_data.id]
 
 ''' 
 creates probability files (.prob files). Probabilities are computed from graph embedding.
 '''
-def compute_probabilities(state_file, destination_dir="./probabilities/", probability_vector_size = 10, users_batch_number=1, total_user_batches=1, event2users=None):
+def compute_probabilities(state_file, embedding_path, users_batch_number=1, total_user_batches=1, event2users=None):
     initial_state_meta_data = read_state_file(state_file)
-    embeddings_data = initial_state_meta_data["embedding_files"]
+    embedding_path = embedding_path #initial_state_meta_data["embedding_path"]
     UsimId2strId = load_id_dictionary(initial_state_meta_data["users_ids"], isSimId2strId=True)
     strId2RsimId = load_id_dictionary(initial_state_meta_data["repos_ids"], isSimId2strId=False)
     total_number_of_agents = len(UsimId2strId)
@@ -223,39 +219,42 @@ def compute_probabilities(state_file, destination_dir="./probabilities/", probab
     def _time(index, start_time):
         if index % 1000 == 0:
             end_time = time.time()
-            print "Agent ", index, " out of ", total_number_of_agents, " : ", 100.0 * float(index) / float(
+            print "Agent ", index, " out of ", total_number_of_agents, " : ",           100.0 * float(index) / float(
                 total_number_of_agents), "%, ", " time:", (end_time - start_time)
             start_time = time.time()
         return start_time
 
-    users_neighborhood_size = pickle.load(open("users_neighborhood_size.pickle", "rb"))
+    users_neighborhood_size = pickle.load(open(initial_state_meta_data["users_neighborhood_sizes_file"], "rb"))
 
-    for event_type, embedding_info in embeddings_data.iteritems():
-        all_probabilities = {}
-        calculator = EmbeddingCalculator(embedding_info["file_name"], embedding_info["dictionary"], UsimId2strId, strId2RsimId)
-        start_time = time.time()
-        if event2users is None:
-            for index, agent_sim_id in enumerate(UsimId2strId.iterkeys()):
-                probability_vector_size = users_neighborhood_size[agent_sim_id] if agent_sim_id in users_neighborhood_size else 10
-                all_probabilities[agent_sim_id] = calculator.calculate_probabilities(agent_sim_id, probability_vector_size)
-                start_time = _time(index, start_time)
-        else:
-            total_number_of_agents = len(event2users[event_type])
-            min_index = (users_batch_number - 1) * int(total_number_of_agents / total_number_of_batches)
-            max_index = users_batch_number * int(total_number_of_agents / total_number_of_batches) if users_batch_number != total_user_batches else total_number_of_agents
-            #print "min: ", min_index, ", max : ", max_index, " total agents: ", total_number_of_agents
-            for index in range(min_index, max_index, 1):
-                user_id = event2users[event_type][index]
-                probability_vector_size = users_neighborhood_size[user_id] if user_id in users_neighborhood_size else 10
-                all_probabilities[user_id] = calculator.calculate_probabilities(user_id, probability_vector_size)
-                start_time = _time(index, start_time)
+    for event_type in event_types:
+        emb_file_path = str(embedding_path) + event_type + ".emb"
+        if os.path.isfile(emb_file_path):
+            id_file_path = str(embedding_path) + event_type + "_nodeID_gf.pickle"
+            all_probabilities = {}
+            calculator = EmbeddingCalculator(emb_file_path, id_file_path, UsimId2strId, strId2RsimId)
+            start_time = time.time()
+            if event2users is None:
+                for index, agent_sim_id in enumerate(UsimId2strId.iterkeys()):
+                    probability_vector_size = users_neighborhood_size[agent_sim_id] if agent_sim_id in users_neighborhood_size else 10
+                    all_probabilities[agent_sim_id] = calculator.calculate_probabilities(agent_sim_id, probability_vector_size)
+                    start_time = _time(index, start_time)
+            else:
+                total_number_of_agents = len(event2users[event_type])
+                min_index = (users_batch_number - 1) * int(total_number_of_agents / total_number_of_batches)
+                max_index = users_batch_number * int(total_number_of_agents / total_number_of_batches) if users_batch_number != total_user_batches else total_number_of_agents
+                #print "min: ", min_index, ", max : ", max_index, " total agents: ", total_number_of_agents
+                for index in range(min_index, max_index, 1):
+                    user_id = event2users[event_type][index]
+                    probability_vector_size = users_neighborhood_size[user_id] if user_id in users_neighborhood_size else 10
+                    all_probabilities[user_id] = calculator.calculate_probabilities(user_id, probability_vector_size)
+                    start_time = _time(index, start_time)
 
-        if total_user_batches > 1:
-            output_file = open(destination_dir + event_type + "_" + str(users_batch_number) + ".prob", 'wb')
-        else:
-            output_file = open(destination_dir + event_type + ".prob", 'wb')
-        pickle.dump(all_probabilities, output_file, protocol=2)
-        output_file.close()
+            if total_user_batches > 1:
+                output_file = open(str(embedding_path) + event_type + "_" + str(users_batch_number) + ".prob", 'wb')
+            else:
+                output_file = open(str(embedding_path) + event_type + ".prob", 'wb')
+            pickle.dump(all_probabilities, output_file, protocol=2)
+            output_file.close()
 
 
 def populate_event_rate(agents, model_file):
@@ -291,7 +290,6 @@ if __name__ == "__main__":
         cmd = raw_input(
             "Press q to exit loader\n\t"
             "l to load objects from profiles file and load them into memory\n\t"
-            "e to load embedding into memory\n\t"
             "p to create probabilities file\n\t"
             "g to gather (merge) probability files\n\t"
             "s to load state file\n")
@@ -307,24 +305,15 @@ if __name__ == "__main__":
             print "Reading state file ..."
             n_users, n_repos = read_state_file(filename)
             print "users: ", n_users, ", repos: ", n_repos
-        elif cmd == "e":
-            filename = sys.argv[1]
-            print "Reading embedding file ..."
-            embedding_ids_file = sys.argv[2]
-            sim_ids_file = sys.argv[3]
-            embedding_calculator = EmbeddingCalculator(filename, embedding_ids_file, sim_ids_file)
-            prob = embedding_calculator.calculate_probabilities(10)
-            print "done: ", prob
         elif cmd == "p":
             state_filename = sys.argv[1]
-            events_filename = sys.argv[2]
-            batch_number = int(sys.argv[3])
-            total_number_of_batches = int(sys.argv[4])
+            embedding_path = sys.argv[2]
+            events_filename = sys.argv[3]
+            batch_number = int(sys.argv[4])
+            total_number_of_batches = int(sys.argv[5])
             event2users = collect_unique_user_event_pairs(events_filename, load_id_dictionary(read_state_file(state_filename)["users_ids"], isSimId2strId=False))
             print "computing probabilities from embedding ..."
-            compute_probabilities(state_file=state_filename, destination_dir="./probabilities/",
-                                  probability_vector_size = 10,
-                                  users_batch_number=batch_number, total_user_batches=total_number_of_batches,
+            compute_probabilities(state_file=state_filename, embedding_path=embedding_path, users_batch_number=batch_number, total_user_batches=total_number_of_batches,
                                   event2users=event2users)
         elif cmd == "g":
             src_dir = sys.argv[1] # must have '/' in the end
